Add helpers usable from AHK right now
Instead of building the full replacement for autohotkey,
we build the python scripts that can be usable by AHK right now.

take_screenshot uses Pillow to take the screenshot and write
it directly to the intended location. This is better because:

- We don't use screenshot key that clutters "recent files"
- We don't use screenshot key that tampers with clipboard (momentarily)
- We don't have to manually move the file after screenshot key writes it
  in one location

get_windows uses pywinctl to capture windows. Honestly, there isn't
any reason this is better than the logic in AHK. But if we are ever
going to migrate to a python-based solution, it's probably helpful
to exercise this logic.