Perf: Avoid recalculation loop on scroll
The FlowBite tooltip component would sneak in a w-fit
which would result in the summary cards to try to fit
to content.

Since we only load images when they scroll into view,
the w-fit would cause each card to resize as the image would
load which would cause all other cards to recalculate.

This felt like it triggered a recursive loop. Used firefox profiling
to see the difference.

Before spend 3s where the main thread was busy in an 18s recording.
In the next, we spent 2ms busy in a 9s recording. 1000x improvement!