Read process data from sqlite
Main motivation was to store/read "frequency" with each
screenshot so that we can have dynamic frequency if we want.

As part oft of that, also did a minor structure change so some data
is not repeated for all processes in a snapshot. Instead they
go into a Snapshot object.